requirements.txt
nose==1.3.0
paramiko==1.12.0
proboscis==1.2.6.0
PyYAML==3.10
selenium==2.35.0
